feat: types and improve structure
