+++ .changeset/mighty-points-return.md
@@ -1,0 +1,5 @@
+---
+"astro-page-insight": patch
+---
+
+Fixes types (hopefully) and improve structure
