+++ package/src/astro-page-insight.ts
@@ -3,24 +3,31 @@ import {
 	defineIntegration,
 	defineOptions,
 } from "astro-integration-kit";
-import { type Options, optionsSchema } from "../types/server";
-import { organizeLHResult, startLH } from "./server";
+import { corePlugins } from "astro-integration-kit/plugins";
+import { type Options, optionsSchema } from "./types.js";
+import { organizeLHResult, startLH } from "./server.js";
 
 export default defineIntegration({
 	name: "astro-page-insight",
+	plugins: [...corePlugins],
 	options: defineOptions<Options>({ weight: 0, breakPoint: 768 }),
 	setup({ options: inputOptions }) {
+		const { weight = 0, breakPoint = 768 } = optionsSchema.parse(inputOptions);
+		const { resolve } = createResolver(import.meta.url);
+
 		return {
-			"astro:config:setup": ({ addDevToolbarApp, command }) => {
+			"astro:config:setup": ({
+				addDevToolbarApp,
+				command,
+				watchIntegration,
+			}) => {
+				watchIntegration(resolve());
+
 				if (command === "dev") {
-					const { resolve } = createResolver(import.meta.url);
-					addDevToolbarApp(resolve("./astro-page-insight-toolbar.ts"));
+					addDevToolbarApp(resolve("./plugin.ts"));
 				}
 			},
-			"astro:server:setup": async ({ server }) => {
-				const { weight = 0, breakPoint = 768 } =
-					optionsSchema.parse(inputOptions);
-
+			"astro:server:setup": async ({ server, logger }) => {
 				server.ws.on(
 					"astro-dev-toolbar:astro-page-insight-app:run-lighthouse",
 					async ({ width, height, url }) => {
@@ -44,6 +51,8 @@ export default defineIntegration({
 								);
 							}
 						} catch (error) {
+							logger.error("Something went wrong");
+							console.error(error);
 							server.ws.send(
 								"astro-dev-toolbar:astro-page-insight-app:on-error",
 								"Something went wrong\nPlease try again",
