+++ package/src/constants.ts
@@ -1,4 +1,4 @@
-export const categories = [
+export const CATEGORIES = [
 	"accessibility",
 	"best-practices",
 	"performance",
