+++ package/src/index.ts
@@ -1,3 +1,3 @@
-import astroPageInsight from "./astro-page-insight";
+import astroPageInsight from "./astro-page-insight.js";
 
 export default astroPageInsight;
