--- package/types/client.ts
@@ -1,6 +1,0 @@
-export type PositionType = {
-	top: number;
-	left: number;
-	width: number;
-	height: number;
-};
